apps/01 - IN PROGRESS/python-app/requirements.txt
Flask
mysql-connector-python
python-memcached
